Debian: Create privilege separation directory
On Debian-based system, the privilege separation directory is
`/run/sshd`. Since `/run` is a tmpfs filesystem, the init script is
responsible for creating the directory when starting `sshd`. However,
this directory might not exist if the `ssh` service has not been
started yet.

This patch ensures that the directory exists before storing the
configuration file. It is required to prevent the `check_cmd` from
failing; specifically, to prevent `sshd` from complaining that the
directory does not exist.